novos exercícios

+++ PythonExercicios/ex010.py
@@ -1,2 +1,2 @@
-n = int(input('Quanto dinheiro você tem na carteira? R$ '))
+n = float(input('Quanto dinheiro você tem na carteira? R$ '))
 print('Você pode comprar {:.2f} doláres'.format(n / 3.27))
